moves language pack importer in adapter
