+++ src/PrestaShopBundle/Controller/Admin/TranslationsController.php
@@ -159,7 +159,7 @@ class TranslationsController extends FrameworkBundleAdminController
             $data = $addUpdateLanguageForm->getData();
             $isoCode = $data['iso_localization_pack'];
 
-            $languagePackImporter = $this->get('prestashop.core.language.pack.import.language_importer');
+            $languagePackImporter = $this->get('prestashop.adapter.language.pack.importer');
             $errors = $languagePackImporter->import($isoCode);
 
             if (empty($errors)) {
