+++ src/PrestaShopBundle/Resources/config/services/core/language.yml
@@ -2,13 +2,6 @@ services:
     _defaults:
         public: true
 
-    prestashop.core.language.pack.import.language_importer:
-        class: 'PrestaShop\PrestaShop\Core\Language\Pack\Import\LanguagePackImporter'
-        arguments:
-            - '@prestashop.adapter.language.language_pack_installer'
-            - '@prestashop.adapter.data_provider.language'
-            - '@prestashop.adapter.cache_clearer'
-
     prestashop.core.language.pack.loader.remote:
         class: 'PrestaShop\PrestaShop\Core\Language\Pack\Loader\RemoteLanguagePackLoader'
         arguments:
